commit ajustes task

+++ src/main/java/br/gov/es/siscap/service/ProjetoService.java
@@ -40,6 +40,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
@@ -176,7 +177,8 @@ public class ProjetoService {
 		logger.info("Buscando projeto com id: {}", id);
 
 		Projeto projeto = Optional.ofNullable(this.buscar(id))
-				.orElseThrow(() -> new IllegalArgumentException("Projeto não encontrado para o ID: " + id));
+				.orElseThrow(() -> new IllegalArgumentException(
+						"Projeto não encontrado ou já foi excluído (ID: %s)".formatted(id)));
 
 		Set<ProjetoPessoa> projetoPessoaSet = projetoPessoaService.buscarPorProjeto(projeto);
 
@@ -213,7 +215,7 @@ public class ProjetoService {
 				guidSUBEO,
 				guidSUBCAP);
 
-		ProjetoDto projetoDtoRetorno = new ProjetoDto(projeto, valorDto, rateio,
+		return new ProjetoDto(projeto, valorDto, rateio,
 				this.buscarIdResponsavelProponente(projetoPessoaSet),
 				this.buscarEquipeElaboracao(projetoPessoaSet),
 				this.buscarSubResponsavelProponente(projetoPessoaSet),
@@ -231,10 +233,8 @@ public class ProjetoService {
 				this.buscarParecer(parecerProjeto),
 				lotacaoUsuario.getValue(),
 				projeto.getProjetoParecerSet().stream().map(ProjetoParecerDto::new).toList(),
-				this.buscarNomeProponente(projetoPessoaSet),
+				projeto.getPessoa().getNome(),
 				projeto.getHistoricoStatus().stream().map(StatusProjetoDto::new).toList());
-
-		return projetoDtoRetorno;
 
 	}
 
@@ -336,11 +336,7 @@ public class ProjetoService {
 				this.enviarEmailGestorAvaliarDic(projeto.getId(), subResponsavelProponente, nomeProponente);
 
 			}
-		} catch (UnsupportedEncodingException e) {
-			logger.error(e.getMessage());
-		} catch (MessagingException e) {
-			logger.error(e.getMessage());
-		} catch (IOException e) {
+		} catch (IOException | MessagingException e) {
 			logger.error(e.getMessage());
 		}
 
@@ -396,7 +392,7 @@ public class ProjetoService {
 
 		// forçar a atualizacao do SUB do responsavel novo
 		projetoPessoaSet.stream()
-				.filter(p -> p.getPessoa().getId().equals(form.idResponsavelProponente())
+				.filter(p -> p.getPessoa().getId() == form.idResponsavelProponente()
 						&& p.getTipoPapel().getId().equals(TipoPapelEnum.RESPONSAVEL_PROPONENTE.getValue()))
 				.findFirst()
 				.ifPresent(p -> {
@@ -417,9 +413,9 @@ public class ProjetoService {
 		List<ProjetoAcaoDto> projetoAcoesDto = form.acoesProjeto();
 		Set<ProjetoAcao> projetoAcoesSet = projetoAcaoService.atualizar(projetoResult, projetoAcoesDto, rascunho);
 
-		String subResponsavelProponente = this.buscarSubResponsavelProponente(projetoPessoaSet);
+		String subResponsavelProponente = projeto.getPessoa().getSub();
 
-		String nomeProponente = this.buscarNomeProponente(projetoPessoaSet);
+		String nomeProponente = projeto.getPessoa().getNome();
 
 		ProjetoParecerDto projetoParecerDto;
 		ProjetoParecer projetoParecer = null;
@@ -512,7 +508,7 @@ public class ProjetoService {
 
 		} else {
 
-			logger.info("Fazer exclusao lógica pois status do DIC esta fora do tratado.", id);
+			logger.info("Fazer exclusao lógica pois status do DIC %s esta fora do tratado.", id);
 
 			this.exclusaoLogica(projeto);
 
@@ -524,7 +520,6 @@ public class ProjetoService {
 
 	}
 
-	@Transactional
 	private void exclusaoLogica(Projeto projeto) {
 
 		projeto.apagarProjeto();
@@ -543,7 +538,6 @@ public class ProjetoService {
 
 	}
 
-	@Transactional
 	private void exclusaoFisica(Projeto projeto) {
 
 		projetoPessoaService.excluirFisicamentePorProjeto(projeto);
@@ -556,13 +550,16 @@ public class ProjetoService {
 
 		projetoParecerService.excluirFisicamentePorProjeto(projeto);
 
+		if (projeto == null) {
+			throw new ValidacaoSiscapException(List.of("Projeto não encontrado para efetivar a exclusao física."));
+		}
+
 		repository.saveAndFlush(projeto);
 
 		repository.deleteFisico(projeto.getId());
 
 	}
 
-	@Transactional
 	public void registrarMotivoArquivamentoProjeto(Long id, String codigoMotivoArquivamento, String justificativa) {
 
 		List<String> erros = new ArrayList<>();
@@ -584,7 +581,6 @@ public class ProjetoService {
 
 	}
 
-	@Transactional
 	public void inserirComplementacoesSeremRealizadasDIC(Projeto projeto,
 			List<ProjetoCamposComplementacaoDto> complementos) {
 
@@ -618,8 +614,9 @@ public class ProjetoService {
 		List<String> erros = new ArrayList<>();
 
 		if (justificativa == null || justificativa.isEmpty() || justificativa.isBlank()) {
-			erros.add("Erro ao enviar solicitação de revisão do projeto id " + id
-					+ " justificativa não presente no pedido de envio.");
+			erros.add(
+					"Erro ao enviar solicitação de revisão do projeto id %s justificativa não presente no pedido de envio."
+							.formatted(id));
 			throw new ValidacaoSiscapException(erros);
 		}
 
@@ -627,19 +624,20 @@ public class ProjetoService {
 
 		Optional<Pessoa> proponenteProjeto = projeto.getProjetoPessoaSet()
 				.stream()
-				.filter(_pessoa -> _pessoa.isProponente())
+				.filter(ProjetoPessoa::isProponente)
 				.findFirst()
-				.map(proponente -> proponente.getPessoa());
+				.map(ProjetoPessoa::getPessoa)
+				.or(() -> Optional.ofNullable(projeto.getPessoa()));
 
 		Optional<Pessoa> responsavelProponenteProjeto = projeto.getProjetoPessoaSet()
 				.stream()
-				.filter(_pessoa -> _pessoa.isResponsavelProponente())
+				.filter(ProjetoPessoa::isResponsavelProponente)
 				.findFirst()
-				.map(proponente -> proponente.getPessoa());
+				.map(ProjetoPessoa::getPessoa);
 
 		if (proponenteProjeto.isPresent() && responsavelProponenteProjeto.isPresent()) {
 
-			List<String> emailsInteressadosList = new ArrayList<String>();
+			List<String> emailsInteressadosList = new ArrayList<>();
 
 			emailsInteressadosList.add(proponenteProjeto.get().getEmail());
 
@@ -658,25 +656,22 @@ public class ProjetoService {
 					projeto.alterarStatus(StatusProjetoEnum.EM_ELABORACAO.getValue(), pessoa);
 
 				} else {
-					erros.add("Erro ao enviar solicitação de revisão do projeto id " + id);
+					erros.add("Erro ao enviar solicitação de revisão do projeto id %s".formatted(id));
 				}
 
-			} catch (UnsupportedEncodingException e) {
-				logger.error(e.getMessage());
-			} catch (MessagingException e) {
+			} catch (UnsupportedEncodingException | MessagingException e) {
 				logger.error(e.getMessage());
 			}
 
 		} else {
-			erros.add("Não foi possível fazer o envio pois o proponente não foi encontrado - projeto id " + id);
+			erros.add("Não foi possível fazer o envio pois o proponente não foi encontrado - projeto id %s"
+					.formatted(id));
 		}
 
 		if (!erros.isEmpty()) {
 			erros.forEach(logger::error);
 			throw new ValidacaoSiscapException(erros);
 		}
-
-		return;
 
 	}
 
@@ -697,20 +692,21 @@ public class ProjetoService {
 
 		Optional<Pessoa> proponenteProjeto = projeto.getProjetoPessoaSet()
 				.stream()
-				.filter(_pessoa -> _pessoa.isProponente())
+				.filter(ProjetoPessoa::isProponente)
 				.findFirst()
-				.map(proponente -> proponente.getPessoa());
+				.map(ProjetoPessoa::getPessoa)
+				.or(() -> Optional.ofNullable(projeto.getPessoa()));
 
 		Optional<Pessoa> responsavelProponenteProjeto = projeto.getProjetoPessoaSet()
 				.stream()
-				.filter(_pessoa -> _pessoa.isResponsavelProponente())
+				.filter(ProjetoPessoa::isResponsavelProponente)
 				.findFirst()
-				.map(proponente -> proponente.getPessoa());
+				.map(ProjetoPessoa::getPessoa);
 
 		if (proponenteProjeto.isPresent() && responsavelProponenteProjeto.isPresent()) {
 
 			boolean confirmacaoEnvioEmail;
-			List<String> emailsInteressadosList = new ArrayList<String>();
+			List<String> emailsInteressadosList = new ArrayList<>();
 			emailsInteressadosList.add(proponenteProjeto.get().getEmail());
 
 			try {
@@ -723,17 +719,15 @@ public class ProjetoService {
 
 				if (confirmacaoEnvioEmail) {
 					logger.info(
-							"Email aviso solicitação de complementação do projeto enviado com sucesso para o projeto id "
-									+ id);
+							"Email aviso solicitação de complementação do projeto enviado com sucesso para o projeto id {}",
+							id);
 					projeto.alterarStatus(StatusProjetoEnum.COMPLEMETACAO.getValue(), pessoa);
 					this.inserirComplementacoesSeremRealizadasDIC(projeto, complementos);
 				} else {
 					erros.add("Erro ao enviar aviso para complementação do projeto id " + id);
 				}
 
-			} catch (UnsupportedEncodingException e) {
-				logger.error(e.getMessage());
-			} catch (MessagingException e) {
+			} catch (UnsupportedEncodingException | MessagingException e) {
 				logger.error(e.getMessage());
 			}
 
@@ -900,7 +894,7 @@ public class ProjetoService {
 						projeto.getId());
 
 				if (confirmacaoEnvioEmail) {
-					logger.info("Email aviso arquivamento projeto enviado com sucesso do projeto id " + id);
+					logger.info("Email aviso arquivamento projeto enviado com sucesso do projeto id {}", id);
 					projeto.alterarStatus(StatusProjetoEnum.ARQUIVADO.getValue(), pessoa);
 					projeto.finalizarStatusAtual(pessoa);
 					this.registrarMotivoArquivamentoProjeto(id, codigoMotivoArquivamento, justificativa);
@@ -962,7 +956,8 @@ public class ProjetoService {
 			Projeto projeto = this.buscar(idProjetoProposto);
 			projeto.setPrograma(programa);
 			repository.saveAndFlush(projeto);
-			if (projetoPropostoSet.stream().noneMatch(projetoSet -> projetoSet.getId().equals(projeto.getId()))) {
+			if (projetoPropostoSet.stream()
+					.noneMatch(projetoSet -> Objects.equals(projetoSet.getId(), projeto.getId()))) {
 				this.enviarAvisoEquipeElaboracaoDicVinculadoPrograma(projeto.getId());
 			}
 		});
@@ -1154,7 +1149,6 @@ public class ProjetoService {
 
 	}
 
-	@Transactional
 	public boolean enviarEmailGestorAvaliarDic(Long idProjeto, String subResponsavelProponente, String nomeProponente)
 			throws MessagingException, UnsupportedEncodingException {
 
@@ -1214,7 +1208,6 @@ public class ProjetoService {
 
 	}
 
-	@Transactional
 	public boolean enviarEmailPareceresEstrategicoOrcamentario(Long idProjeto, String subResponsavelProponente,
 			String nomeProponente) throws MessagingException, UnsupportedEncodingException {
